backend/requirements.txt
todoist-api-python
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-genai
lancedb>=0.24.0
pymongo
motor
python-dotenv
fastapi
python-multipart
uvicorn[standard]
gunicorn
python-socketio
redis
pypdf
pypdf2
firebase-admin
bcrypt
structlog
logtail-python
asgi-correlation-id
pandas
numpy

# Testing
pytest
pytest-asyncio
httpx